requirements.txt
#Danh sách thư viện

python-dotenv
